+++ routers/api/v1/notify/user.go
@@ -11,6 +11,7 @@ import (
 
 	"code.gitea.io/gitea/models"
 	"code.gitea.io/gitea/modules/context"
+	"code.gitea.io/gitea/modules/convert"
 	"code.gitea.io/gitea/routers/api/v1/utils"
 )
 
@@ -87,7 +88,7 @@ func ListNotifications(ctx *context.APIContext) {
 		return
 	}
 
-	ctx.JSON(http.StatusOK, nl.APIFormat())
+	ctx.JSON(http.StatusOK, convert.ToNotifications(nl))
 }
 
 // ReadNotifications mark notification threads as read, unread, or pinned
